test: Move complex parameter test to the RemoteInjectionTest class and make the tests sequential
